Rename CPUInfo socketId to physicalId
Though both names are equivalent from a pure practical view, physicalId is
better than socketId in the sense that it aligns with the name of the source of
its value on /proc/cpuinfo and with the nomenclature used through the kernel
source.

Compatibility with profiles that uses the socketId attribute name is retained.