Fix window association with the application desktop file under Wayland
Wayland DEs use the desktopFileName Wayland property in order to associate
windows with the appropriate .desktop file.

Qt5 sets this to the binary name by default which does not match the previous
name of the .desktop file. Fix this by directly calling `setDesktopFileName`
with the name of the .desktop file.

This should fix Wayland DEs while not breaking X11 ones. Qt5 on X11 sets the
resourceClass to the binary name, but this is fine since windows can be matched
by the `StartupWMClass` property in the .desktop file which is already set to
the binary name.

Also while we're at it fix the .desktop file and appdata.xml to use
`org.corectrl.CoreCtrl` instead of `org.corectrl.corectrl`. This better follows
the FreeDesktop specifications.