Fix compilation with Linux 6.6 API headers
Due to some changes in cn_proc.h in Linux 6.6 headers, including and using it
mixed with C++ code is no longer posible without specifying that the code is C
code and must be compiled with a C compiler.

Isolate all process connector related code into its own C file, which is now
compiled using a C compiler, and adapt it to adhere to the new changes.

Fixes #393.