PMPowerCap: use power1_cap_default value when available
On recent kernel versions, some cards lacks support of resetting the current
value by writing a 0 into power1_cap. When this is done during corectrl
initialization, a warning message is issued to the kernel log, and the
application will use the pre-init value as the default one.

Use power1_cap_default value if it's available and fallback to the old method
when it's not.

Fixes #422.