Generalize Input type of SysFSDataSource
This change introduces a new Input template parameter to SysFSDataSource.

The Input parameter has the type SysFSDataType and it allows to transform the
whole source data to a Output type data by using the transformation function
supplied at construction time and the type std::vector<string> is used as the
Input template parameter.