Make the fixed frequency control a legacy control
The fixed frequency control load the available states from the pp_dpm_* files at
application startup. This was fine when this control was originally implemented
as the states were the same no matter the current power state. However, this is
no longer the case for newer drivers and GPUs: pp_dpm_* files will have
different states depending on the current power state, which turns them a
non-reliable source of data.

Therefore, the usage of the fixed frequency control must be limited to older
kernels and cannot replace the profile based fixed state control (reverting part
of the intention of 17b2824).