Invert application components creation
Some of Qt components depends on an already constructed and initialized
QApplication to work properly (like QCommandLineParser and QStandardPaths).

The ability to parse command line arguments is quite useful to implement
commands that might affect the creation of the core components.

With the component creation inversion the application is in charge of creating
the rest of the components once the main instance of QApplication has been
created, allowing the use of more Qt functionality earlier in the construction
process.