Fix wrong profile settings being applied when a manual profile is active
When a manual profile is renamed while is active or an automatic profile becomes
active, the settings from the penultimate active profile are used instead of the
last one.

While recreating the profile views, the last remaining profile view
should be used as the base view. It doesn't matter whether a manual profile exits
or not, as the affected profile views are removed from the profile views just
before the recreation step.